requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic-settings==2.5.2
apscheduler==3.10.4
anthropic==0.40.0
httpx==0.27.2
playwright==1.48.0
python-multipart==0.0.12
